+++ codedungeon.Tester/src/cdTester.js
@@ -1,0 +1,5 @@
+// @flow
+
+export async function testerHeartbeat(): Promise<void> {
+  console.log('testerHeartbeat')
+}